Refactor iOS local database core lifecycle
